refactor: 최근 게시물 api 로직 위치 변경
